testMap: recliquer sur une vitrine la met en arrière-plan

+++ HotPoints_Projects/TestLocation/app/src/main/java/vitrine/com/testmap/Vitrine.java
@@ -30,7 +30,7 @@ public class Vitrine {
         return latLng;
     }
 
-    public double getRadius() {
+    public int getRadius() {
         return radius;
     }
 }
